+++ cogs/general.py
@@ -19,10 +19,13 @@ class General(commands.Cog):
         )
 
     #Commands
+    #Pings user and checks client latency to the server. Used for Bot-end testing.
     @commands.command()
     async def ping(self, ctx):
         await ctx.send(f'Pong! {round(self.client.latency * 1000)}ms')
     
+    #Display user profile using embeds - no additional user permissions required
+    #other variations that this command can be called are user or avatar
     @commands.command(aliases=['avatar', 'user'])
     async def userinfo(self, ctx, member: discord.Member = None):
         roles = [role for role in ctx.author.roles[1:]]
@@ -40,6 +43,8 @@ class General(commands.Cog):
         embed.add_field(name=f"Allocated roles: ({len(roles)})", value="".join([role.mention for role in roles]))
         await ctx.send(embed=embed)
     
+    #Display server profile via embeds.
+    #other variations that said command can be called is server_profile
     @commands.command(aliases=['server_profile'])
     async def serverinfo(self, ctx):
 
@@ -48,40 +53,8 @@ class General(commands.Cog):
         embed.add_field(name="Number of members:", value=ctx.guild.member_count, inline=False)
         embed.add_field(name="Server created:", value=ctx.guild.created_at.strftime("%a, %d %B %Y, %I:%M %p UTC"), inline=False)
         await ctx.send(embed=embed)
-		
-    @commands.command()
-    async def get_history(self, message):
-        #empty dataframe
-        data = pd.DataFrame(columns=['msg_id', 'content', 'time',
-                                    'author', 'channel'])
 
-        #grab list of channel history in #feedback
-        messages = await self.client.get_channel(845906253326188574).history(limit=200).flatten()
-
-        #loop through each message and save to empty dataframe
-        for msg in messages:
-            data = data.append({'msg_id': msg.id,
-                                    'content': msg.content,
-									'time': msg.created_at,
-									'author': msg.author.name,
-                                    'channel': msg.channel
-                                    }, ignore_index=True)
-        
-        #declare file location + save as csv for analytics
-        file_location= "data.csv"
-        data.to_csv(file_location)
-        
-        await message.channel.send("Data saved. Consult your administrator for data file.")
-
-    @commands.command(aliases=['changecolour'])
-    @commands.has_permissions(manage_roles=True)
-    async def changecolor(self, ctx, role: discord.Role):
-        try:
-            await role.edit(color=discord.Color(random.randint(0x000000, 0xFFFFFF)))
-            await ctx.send("Colour has been changed. Check role to see new colour")
-        except Exception as error:
-            raise(error)
-
+#Add cog of generic functions
 def setup(client):
     client.add_cog(General(client))
