+++ main.py
@@ -1,3 +1,6 @@
+#Before you run
+#Please grab .env file before running. Find this in Discord Developer Portal or contact the bot account's creator
+#The Discord Token is required to authenticate connection between client -> discord
 import os
 import random
 import discord
@@ -9,8 +12,10 @@ load_dotenv()
 TOKEN = os.getenv('DISCORD_TOKEN')
 GUILD = os.getenv('DISCORD_GUILD')
 
+#Create an instance of bot which users access commands by calling !<command>
 client = commands.Bot(command_prefix="!")
 
+#Init bot, display connection success in cmd
 @client.event
 async def on_ready():
     guild = discord.utils.get(client.guilds, name=GUILD)
@@ -20,14 +25,17 @@ async def on_ready():
         f'{guild.name}(id: {guild.id})'
         )
 
+#Load all cogs in /cogs
 @client.command()
 async def load(ctx, extension):
     client.load_extension(f'cogs.{extension}')
 
+#Unload cogs for error testing
 @client.command()
 async def unload(ctx, extension):
     client.unload_extension(f'cogs.{extension}')
 
+#Reload all cog files
 @client.command()
 async def reload(ctx, extension):
     client.unload_extension(f'cogs.{extension}')
@@ -37,4 +45,5 @@ for filename in os.listdir('./cogs'):
     if filename.endswith('.py'):
         client.load_extension(f'cogs.{filename[:-3]}')
 
+#Send bot client token environment variables
 client.run(TOKEN)
